Fix `on()` queue option handling
This commit fixes an issue where the on() function prototype was not
creating its queue with the options passed to it (exclusive, durable,
autoDelete).  The code is modified to create an object that is passed to
the amqp library.  The option contains the properties durable, exclusive
and autoDelete.

Defaults match node-amqp's defaults.

Tested against node-amqp 0.2.4 and RabbitMQ 3.3.5